Eliminate platform differences.

+++ src/native_system.cpp
@@ -11,7 +11,7 @@ REXPORT int RCALL munmap_ffi(void* addr, size_t len) {
   return munmap(addr, len);
 }
 
-REXPORT int RCALL chsize_ffi(int file_handle, long size) {
+REXPORT int RCALL chsize_ffi(int file_handle, int32_t size) {
 #ifdef _MSVC_LANG
   return _chsize(file_handle, size);
 #else
